Refactor swap function
