--- src/koeda/random_swap.py
@@ -1,22 +1,0 @@
-import random
-
-
-def swap_word(new_words):
-    random_idx_1 = random.randint(0, len(new_words) - 1)
-    random_idx_2 = random_idx_1
-    counter = 0
-    while random_idx_2 == random_idx_1:
-        random_idx_2 = random.randint(0, len(new_words) - 1)
-        counter += 1
-        if counter > 3:
-            return new_words
-    new_words[random_idx_1], new_words[random_idx_2] = new_words[random_idx_2], new_words[random_idx_1]
-    return new_words
-
-def random_swap(words, n):
-    new_words = words.copy()
-    for _ in range(n):
-        new_words = swap_word(new_words)
-    return new_words
-
-
